fix trang home, giao dien dep hon

+++ database/migrations/2025_07_07_145028_add_sale_percent_to_product_variants_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up()
     {
         Schema::table('product_variants', function (Blueprint $table) {
-            $table->decimal('sale_percent', 5, 2)->default(0)->after('original_price');
+            $table->integer('sale_percent')->default(0)->after('original_price');
         });
     }
